# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/tax_templates:
    get:
      tags:
        - TaxTemplates
      summary: List TaxTemplates
      description: Get a paginated list of TaxTemplates with optional filtering and sorting
      operationId: listTaxTemplates
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaxTemplateList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - TaxTemplates
      summary: Create TaxTemplate
      description: Create a new TaxTemplate
      operationId: createTaxTemplate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaxTemplateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaxTemplate'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/tax_template_rows:
    get:
      tags:
        - TaxTemplateRows
      summary: List TaxTemplateRows
      description: Get a paginated list of TaxTemplateRows with optional filtering and sorting
      operationId: listTaxTemplateRows
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaxTemplateRowList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - TaxTemplateRows
      summary: Create TaxTemplateRow
      description: Create a new TaxTemplateRow
      operationId: createTaxTemplateRow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaxTemplateRowRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaxTemplateRow'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    TaxTemplate:
      type: object
      required:
        - id
        - code
        - name
        - template_type
        - is_inclusive
        - tax_exigibility
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
          maxLength: 40
        name:
          type: string
          maxLength: 150
        template_type:
          type: string
        tax_category_id:
          nullable: true
          type: string
          format: uuid
        is_inclusive:
          type: boolean
        tax_exigibility:
          type: string
        cash_basis_transition_account_id:
          nullable: true
          type: string
          format: uuid
        status:
          type: string
        metadata:
          type: object
...
    CreateTaxTemplateRequest:
      type: object
      required:
        - code
        - name
        - template_type
        - is_inclusive
        - tax_exigibility
        - status
        - metadata
      properties:
        code:
          type: string
          maxLength: 40
        name:
          type: string
          maxLength: 150
        template_type:
          type: string
        tax_category_id:
          nullable: true
          type: string
          format: uuid
        is_inclusive:
          type: boolean
        tax_exigibility:
          type: string
        cash_basis_transition_account_id:
          nullable: true
          type: string
          format: uuid
        status:
          type: string
        metadata:
          type: object
...
    TaxTemplateList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TaxTemplate'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    TaxTemplateRow:
      type: object
      required:
        - id
        - template_id
        - charge_type
        - rate
        - is_withholding
        - effective_from
        - sort_order
        - metadata
      properties:
        id:
          type: string
          format: uuid
        template_id:
          type: string
          format: uuid
        charge_type:
          type: string
        rate:
          type: string
        account_id:
          nullable: true
          type: string
          format: uuid
        is_withholding:
          type: boolean
        effective_from:
          type: string
          format: date
        effective_to:
          nullable: true
          type: string
          format: date
        sort_order:
          type: integer
          format: int32
        description:
          nullable: true
          type: string
          maxLength: 150
        metadata:
          type: object
...
    CreateTaxTemplateRowRequest:
      type: object
      required:
        - template_id
        - charge_type
        - rate
        - is_withholding
        - effective_from
        - sort_order
        - metadata
      properties:
        template_id:
          type: string
          format: uuid
        charge_type:
          type: string
        rate:
          type: string
        account_id:
          nullable: true
          type: string
          format: uuid
        is_withholding:
          type: boolean
        effective_from:
          type: string
          format: date
        effective_to:
          nullable: true
          type: string
          format: date
        sort_order:
          type: integer
          format: int32
        description:
          nullable: true
          type: string
          maxLength: 150
        metadata:
          type: object
...
    TaxTemplateRowList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TaxTemplateRow'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []